refactor: move Playwright controller to the project root
- extract Playwright controller
- remove unnecessary jasmine plugins
