--- e2e-app/playwright/helpers/jasmine-ts.js
@@ -1,8 +1,0 @@
-const path = require("path");
-
-const config = {
-  // eslint-disable-next-line global-require
-  project: path.join(__dirname, "../../tsconfig.spec.json")
-};
-// console.log('jasmine-ts', config);
-require("ts-node").register(config);
