+++ package.json
@@ -14,9 +14,9 @@
     "build": "yarn ngb:build-ivy && yarn ngb:build && yarn demo:build",
     "test": "yarn check-format && yarn ngb:lint && yarn ngb:test && yarn schematics:test",
     "tdd": "yarn ngb:tdd",
-    "playwright": "wait-on http://localhost:4200 && jasmine JASMINE_CONFIG_PATH=e2e-app/playwright/jasmine.json",
-    "playwright:firefox": "wait-on http://localhost:4200 && jasmine JASMINE_CONFIG_PATH=e2e-app/playwright/jasmine.json BROWSER=firefox",
-    "playwright:webkit": "wait-on http://localhost:4200 && jasmine JASMINE_CONFIG_PATH=e2e-app/playwright/jasmine.json BROWSER=webkit",
+    "playwright": "wait-on http://localhost:4200 && ts-node --project e2e-app/tsconfig.spec.json node_modules/jasmine/bin/jasmine --config=e2e-app/jasmine.json",
+    "playwright:firefox": "wait-on http://localhost:4200 && ts-node --project e2e-app/tsconfig.spec.json node_modules/jasmine/bin/jasmine --config=e2e-app/jasmine.json BROWSER=firefox",
+    "playwright:webkit": "wait-on http://localhost:4200 && ts-node --project e2e-app/tsconfig.spec.json node_modules/jasmine/bin/jasmine --config=e2e-app/jasmine.json BROWSER=webkit",
     "e2e": "yarn e2e-app:lint && yarn e2e:playwright && yarn demo:e2e",
     "e2e:playwright": "concurrently --success first --kill-others \"yarn e2e-app:serve\" \"yarn playwright && yarn playwright:firefox && yarn playwright:webkit\"",
     "demo": "yarn demo:docs && yarn demo:stackblitzes && yarn demo:serve",
